refactor: reorganize project structure into clear package hierarchy
