--- src/main/java/org/example/model/Grass.java
@@ -1,14 +1,0 @@
-package org.example.model;
-
-import org.example.map.Cell;
-import org.example.map.IslandMap;
-import org.example.statistics.StatisticTracker;
-import org.example.util.RandomUtil;
-
-import java.util.ArrayList;
-
-public class Grass extends Plant {
-   public Grass() {
-       super(15, 1, 200);
-   }
-}
