+++ src/main/java/org/example/model/animals/Herbivore.java
@@ -1,4 +1,4 @@
-package org.example.model;
+package org.example.model.animals;
 
 public abstract class Herbivore extends Animal {
     public Herbivore(int id, double foodNeeded, int speed, int maxAmount, double weight) {
