+++ src/main/java/org/example/model/animals/Rabbit.java
@@ -1,0 +1,7 @@
+package org.example.model.animals;
+
+public class Rabbit extends Herbivore {
+    public Rabbit() {
+        super(7, 0.45, 2, 150, 2);
+    }
+}
