+++ src/main/java/org/example/model/plants/Grass.java
@@ -1,0 +1,7 @@
+package org.example.model.plants;
+
+public class Grass extends Plant {
+   public Grass() {
+       super(15, 1, 200);
+   }
+}
