+++ src/main/java/org/example/statistics/StatisticPrinter.java
@@ -2,8 +2,8 @@ package org.example.statistics;
 
 import org.example.map.IslandMap;
 import org.example.map.Cell;
-import org.example.model.Animal;
-import org.example.model.Plant;
+import org.example.model.animals.Animal;
+import org.example.model.plants.Plant;
 
 import java.util.ArrayList;
 import java.util.HashMap;
